Fix critical bugs and enhance Item Database with visual indicators (#1)
* fix: resolve NumPy type compatibility with PIL/Qt6

Python 3.13 + NumPy 2.0+ returns numpy.bool_ and numpy.int64 types that are incompatible with PIL and Qt6, causing runtime crashes with error: 'numpy.bool' object cannot be interpreted as an integer.

Changes:
- Convert boolean mask to uint8 before scipy.ndimage.label()
- Wrap all numeric values with int() for native Python types
- Apply to GetSystemMetrics(), argmax(), and bbox coordinates

Fixes screen capture crashes during OCR processing.

* fix: add Tesseract path detection for development mode

Application was throwing TesseractNotFoundError when running from source code (non-compiled) because it only checked for bundled PyInstaller path.

Changes:
- Added fallback to check local Tesseract-OCR folder in script directory
- Enables running from source without manual PATH configuration

Allows developers to run the app directly without building.

* feat: enhance Item Database with visual indicators and filtering

Improves user experience by showing where items are needed and adding focused filtering options.

New Features:
1. Enhanced Tooltips
   - Show exact quantities needed per quest/hideout/project
   - Example: 'Med Station Lv.2 (10x)' instead of just 'Hideout'
   - Display up to 3 items + count for remaining

2. Colored Border Indicators on TRACK buttons
   - Green = Required for Quests
   - Blue = Required for Hideout
   - Orange = Required for Projects
   - Borders divided proportionally for multiple categories

3. 'Needed For' Filter Dropdown
   - Filter items by Quests/Hideout/Projects
   - Works with existing search and filters
   - Helps focus on specific progression goals

4. Visual Legend Bar
   - Displays color meanings for new users
   - Positioned between filters and item grid

Implementation:
- Added check_item_needed_for() method to determine item requirements
- Updated tooltips to parse quantities from JSON data
- Modified update_track_button_style() for dynamic border colors
- All changes backward compatible with existing progress.json

User Benefits:
- Instant visibility of required item quantities
- Quick identification of progression-critical items
- Ability to focus on specific game objectives

+++ modules/image_processor.py
@@ -11,25 +11,25 @@ class ImageProcessor:
         r, g, b = target_color
         mask = ((np.abs(img_array[:, :, 0] - r) <= tolerance) & 
                 (np.abs(img_array[:, :, 1] - g) <= tolerance) & 
-                (np.abs(img_array[:, :, 2] - b) <= tolerance))
+                (np.abs(img_array[:, :, 2] - b) <= tolerance)).astype(np.uint8)
         labeled_mask, num_features = label(mask)
-        if num_features == 0:
+        if int(num_features) == 0:
             return None
         blob_sizes = np.bincount(labeled_mask.ravel())
         blob_sizes[0] = 0
-        largest_blob_label = blob_sizes.argmax()
+        largest_blob_label = int(blob_sizes.argmax())
         if largest_blob_label == 0:
             return None
-        blob_slices = find_objects(labeled_mask == largest_blob_label)
+        blob_slices = find_objects((labeled_mask == largest_blob_label).astype(np.uint8))
         if not blob_slices:
             return None
         y_slice, x_slice = blob_slices[0]
         padding = 5
         return (
-            max(0, x_slice.start - padding), 
-            max(0, y_slice.start - padding), 
-            min(img.width, x_slice.stop + padding), 
-            min(img.height, y_slice.stop + padding)
+            int(max(0, x_slice.start - padding)), 
+            int(max(0, y_slice.start - padding)), 
+            int(min(img.width, x_slice.stop + padding)), 
+            int(min(img.height, y_slice.stop + padding))
         )
     
     @staticmethod
@@ -37,21 +37,21 @@ class ImageProcessor:
         try:
             if full_screen:
                 # Capture the entire primary screen
-                screen_width = ctypes.windll.user32.GetSystemMetrics(0)
-                screen_height = ctypes.windll.user32.GetSystemMetrics(1)
+                screen_width = int(ctypes.windll.user32.GetSystemMetrics(0))
+                screen_height = int(ctypes.windll.user32.GetSystemMetrics(1))
                 search_bbox = (0, 0, screen_width, screen_height)
             else:
                 # Capture around the cursor
                 pt = ctypes.wintypes.POINT()
                 ctypes.windll.user32.GetCursorPos(ctypes.byref(pt))
-                screen_width = ctypes.windll.user32.GetSystemMetrics(0)
-                screen_height = ctypes.windll.user32.GetSystemMetrics(1)
+                screen_width = int(ctypes.windll.user32.GetSystemMetrics(0))
+                screen_height = int(ctypes.windll.user32.GetSystemMetrics(1))
                 search_width, search_height = 1200, 1200
                 search_bbox = (
-                    max(0, pt.x - search_width // 2),
-                    max(0, pt.y - search_height // 2),
-                    min(screen_width, pt.x + search_width // 2),
-                    min(screen_height, pt.y + search_height // 2)
+                    int(max(0, pt.x - search_width // 2)),
+                    int(max(0, pt.y - search_height // 2)),
+                    int(min(screen_width, pt.x + search_width // 2)),
+                    int(min(screen_height, pt.y + search_height // 2))
                 )
 
             search_img = ImageGrab.grab(bbox=search_bbox)
